Switched from path handling to imagefield

+++ setup/image_downloader.py
@@ -4,41 +4,71 @@ import os
 import requests
 import shutil
 
+
 from zipfile import ZipFile
+from PIL import Image
+import django
+
+django.setup()
+
+from data.models import Paper
+from django.db.models import Q
+
+from io import BytesIO
+from django.core.files.uploadedfile import InMemoryUploadedFile
+from django.core.files.base import ContentFile
+
+def cleaned_doi(paper):
+    return paper.doi.replace("/", "").replace(".", "")
 
 
 if __name__ == "__main__":
-    DOWNLOAD_URL = 'https://dl.dropboxusercontent.com/s/ypzbzlb60b8acmc/pdf_images.zip'
 
-    IMAGES_FOLDER = 'pdf_images'
+    print(Paper.objects.filter(~Q(preview_image__exact='')).count())
 
-    TMP_DIR = 'tmp'
-    TMP_FILENAME = 'images.zip'
+    if Paper.objects.filter(~Q(preview_image__exact='')).count() == 0:
+        DOWNLOAD_URL = 'https://dl.dropboxusercontent.com/s/ypzbzlb60b8acmc/pdf_images.zip'
 
-    FILE_PATH = os.path.join(TMP_DIR, TMP_FILENAME)
-    EXTRACTED_FOLDER_PATH = os.path.join(TMP_DIR, IMAGES_FOLDER)
+        IMAGES_FOLDER = 'pdf_images'
 
-    if os.path.isdir(os.path.join("static", IMAGES_FOLDER)):
-        print(f'Image Folder already present, ending..', os.path.join("static", IMAGES_FOLDER))
-        exit(0)
+        TMP_DIR = 'tmp'
+        TMP_FILENAME = 'images.zip'
 
-    if not os.path.isdir(TMP_DIR):
-        print('Creating tmp directory')
-        os.mkdir(TMP_DIR)
+        FILE_PATH = os.path.join(TMP_DIR, TMP_FILENAME)
+        EXTRACTED_FOLDER_PATH = os.path.join(TMP_DIR, IMAGES_FOLDER)
 
-    print('Downloading zip file.')
-    r = requests.get(DOWNLOAD_URL)
+        if not os.path.isdir(TMP_DIR):
+            print('Creating tmp directory')
+            os.mkdir(TMP_DIR)
 
-    with open(FILE_PATH, 'wb') as f:
-        f.write(r.content)
-    print('Download complete. Extracting..')
+        print('Downloading zip file.')
+        r = requests.get(DOWNLOAD_URL)
 
-    with ZipFile(FILE_PATH, 'r') as z:
-        z.extractall(path=TMP_DIR)
+        with open(FILE_PATH, 'wb') as f:
+            f.write(r.content)
+        print('Download complete. Extracting..')
 
-    print(f'Moving {EXTRACTED_FOLDER_PATH} to {IMAGES_FOLDER}')
+        with ZipFile(FILE_PATH, 'r') as z:
+            z.extractall(path=TMP_DIR)
 
-    shutil.move(EXTRACTED_FOLDER_PATH, os.path.join("static", IMAGES_FOLDER))
+        for paper in Paper.objects.all():
 
-    print(f'Cleaning up: Removing temp directory {TMP_DIR}')
-    shutil.rmtree(TMP_DIR)
+            image = Image.open(os.path.join(EXTRACTED_FOLDER_PATH, cleaned_doi(paper) + ".jpg"))
+
+            buffer = BytesIO()
+            image.save(fp=buffer, format='JPEG')
+
+            pillow_image = ContentFile(buffer.getvalue())
+
+            img_name = cleaned_doi(paper) + ".jpg"
+            paper.preview_image.save(img_name, InMemoryUploadedFile(
+                pillow_image,  # file
+                None,  # field_name
+                img_name,  # file name
+                'image/jpeg',  # content_type
+                pillow_image.tell,  # size
+                None))
+
+            paper.save()
+
+        shutil.rmtree(TMP_DIR)
